+++ src/main/java/homeTry/team/controller/TeamController.java
@@ -3,6 +3,7 @@ package homeTry.team.controller;
 import homeTry.common.annotation.LoginMember;
 import homeTry.member.dto.MemberDTO;
 import homeTry.team.dto.*;
+import homeTry.team.dto.request.TeamCreateRequest;
 import homeTry.team.dto.response.NewTeamFromResponse;
 import homeTry.team.dto.response.RankingResponse;
 import homeTry.team.dto.response.TeamResponse;
