+++ src/main/java/homeTry/team/dto/request/TeamCreateRequest.java
@@ -1,4 +1,4 @@
-package homeTry.team.dto;
+package homeTry.team.dto.request;
 
 import homeTry.tag.dto.TagDTO;
 import jakarta.validation.Valid;
