+++ src/main/java/homeTry/team/service/TeamService.java
@@ -9,6 +9,7 @@ import homeTry.tag.dto.TagDTO;
 import homeTry.tag.model.entity.Tag;
 import homeTry.tag.service.TagService;
 import homeTry.team.dto.*;
+import homeTry.team.dto.request.TeamCreateRequest;
 import homeTry.team.dto.response.NewTeamFromResponse;
 import homeTry.team.dto.response.RankingResponse;
 import homeTry.team.dto.response.TeamResponse;
